Destroy() cleanups in VOREStation code
* Remove remaining hard del()'s in our code.
* Replace deleted() and gcDestroyed with QDELETED macro.
* Fix some strange calls directly to Del() (capital D)
* Make Destroy() return qdel hints.
* Make a few of the Destroy()'s more comprehensive in cleaning up references.
* In edited Destroy() procs, converted to use qdel_null and qdel_null_list macros when possible for pretty code.
* Removed unused variable `sleevecard`

+++ code/modules/nifsoft/nifsoft.dm
@@ -50,11 +50,10 @@
 //Destructor cleans up the software and nif reference
 /datum/nifsoft/Destroy()
 	if(nif)
-		uninstall(nif)
+		uninstall()
 		nif = null
-	qdel(stat_line)
-	stat_line = null
-	..()
+	qdel_null(stat_line)
+	return ..()
 
 //Called when the software is installed in the NIF
 /datum/nifsoft/proc/install()
@@ -67,7 +66,8 @@
 	if(nif)
 		. = nif.uninstall(src)
 		nif = null
-	qdel(src)
+	if(!QDESTROYING(src))
+		qdel(src)
 
 //Called every life() tick on a mob on active implants
 /datum/nifsoft/proc/life(var/mob/living/carbon/human/human)
@@ -126,7 +126,7 @@
 /datum/nifsoft/package/Destroy()
 	software.Cut()
 	software = null
-	..()
+	return ..()
 
 /////////////////
 // A NIFSoft Disk
